% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\docType{data}
\name{bpd}
\alias{bpd}
\title{Persistent dataset to hold blood pressure related data.}
\format{
Data frame
}
\description{
\itemize{
   \item DATE Date column for record
   \item SYS numeric Systolic bp value
   \item DIAS numeric Diastolic bp value
   \item WT numeric Body weight that day in pounds
   \item EX numeric Minutes of exercise in last 24 hours
   \item MDX numeric Blood pressure medicine dosage taken that day in milligrams
 }
}
\keyword{datasets}
